\name{DFX data analysis internal functions}
\alias{dnor}
\alias{radd}
\alias{raddifsad}
\alias{radiald}
\alias{radialdifd}
\alias{sieveDFXdata}
\alias{smap}


\title{Darwin - internal functions}
\description{Internal functions serving DFX data analysis tool}
\details{
 No details.
}
\value{  
	No return value.
}
\author{Pavel Nesladek <Pavel.Nesladek@amtc-dresden.com> }
\keyword{misc}
