\name{readLT}
\alias{readLT}
\title{Lasertec blank inspection data import}
\description{Read Lasertec blank inspection file.}
\usage{
readLT(Fname, CATS=TRUE)
}
\arguments{
	\item{Fname}{inspection report filename placed in working directory, the assigned tool path or specified by fullpath}
	\item{CATS}{parameter decides if defect coordinates will be transformed to CATS coordinates (shifted) -default TRUE. If FALSE, alignment will be kept as in data file. By default, the coordinates units are mm referring to ideal lower left corner of the mask blank. If \bold{CATS=FALSE}, Origin is mask center.}
}
\details{
	Function reads the defect list from inspection report generated by Lasertec blank inspection tool. 
Data from both front and back side groups are joined and additional column \bold{Id} is added, enumerating all defects in scan. In raw data, the defect Ids are not unique. 
Current version of \bold{readLT()} function is looking for data in working directory and in folder defined for the tool data at T drive. In case the file is not found, function stops with error message. 
The defect Id is not unique. As there are two section for front side and back side defects, the defect Id in both groups are overlapping.
Inspection file may contain DISE data - calibrated size measurement of defect size, for predefined defect count/ most frequntly 100. If so, these data sets are read and overwrite the columns \bold{SizeX} and \bold{Size} in resulting data.\\
Function is build on assumption, that defect Ids are steadily increasing in all data sets. If not, the calibrated defect sizes, may be assigned wrongly. For this reason you are strongly adviced against manual modification of the raw data files. 

}

\value{  
	The return value is data frame containing one data row per defect.  The columns of the data frame are following:
	\item{Id}{unique defect indetification Id assigned by \bold{readLT} function.}
	\item{DefNum}{defect index in inspection report}
	\item{Side}{Front / Back Side of the mask}
	\item{X}{X coordinate of the defect according to parameter \bold{CATS}. Units are mm referring to lower left mask corner (default) or mask center \bold{CATS=FALSE)}. In inspection report coordinates are referring to mask center, for consistency reasons and compatibility to DFX data and functions in darwintools the coordinates origin is shifted by default by 76.2mm to the left.}
	\item{Y}{Y coordinate of the defect according to parameter \bold{CATS}. Units are mm referring to lower left mask corner (default) or mask center \bold{CATS=FALSE)}. In inspection report coordinates are referring to mask center, for consistency reasons and compatibility to DFX data and functions in darwintools the coordinates origin is shifted by default by 76.2mm to bottom.}
	\item{Type}{Defect type}
	\item{Size}{Defect size in microns, equal to \bold{SizeX} independently from the fact if calibrated or not. }
	\item{Channel}{Defect }
	\item{Pixel}{Defect size in Pixels}
	\item{Bin}{Defect bin according to size}
}
\author{Pavel Nesladek}
\seealso{readDFX, viewDFXplot, subtractDFX}
\examples{
## Do not run
# data  <- readLT("9A000112.csv", CATS=TRUE)
# viewDFXplot(data, pch=19, cex=0.7, col="red")

## add second inspection data set  from Nu Flare tool
# data2  <- readNF("5000026_282112.txt")
# viewDFXplot(data2, add=TRUE, pch=19, cex=0.4, col="blue")

}
\keyword{misc}
